# openapi/spec.yaml
openapi: "3.0.3"
info:
  title: Eino Demo API
  version: "1.0.0"
  description: Shared API types for Go and TypeScript code generation.
    This spec defines both HTTP endpoint types and WebSocket envelope types.
paths:
...
  /conversations/{id}/messages:
    get:
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        "200":
          description: Message created
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                  message_id:
                    type: string
                  seq:
                    type: integer

  # ── Chat Control ──
  /conversations/{id}/stop:
    post:
      summary: Stop active streaming in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Streaming stopped
  /conversations/{id}/answer:
    post:
      summary: Answer an interrupted human-in-the-loop question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [checkpoint_id, interrupt_id, answer]
              properties:
                checkpoint_id:
                  type: string
                  description: Eino checkpoint ID for resuming the agent run
                interrupt_id:
                  type: string
                  description: Eino interrupt signal ID for targeted resume
                answer:
                  type: string
                  description: User's answer text
      responses:
        "200":
          description: Agent resumed with user's answer
        "404":
          description: Conversation or HITL request not found
        "400":
          description: Invalid request or HITL already answered
  /conversations/{id}/hitl:
    get:
      summary: List pending human-in-the-loop requests for a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of pending HITL requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HumanInTheLoop"
        "404":
          description: Conversation not found
  /conversations/{id}/todos:
    get:
      summary: List todos for a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      summary: Create a todo item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        "201":
          description: Created todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
  /todos/{id}:
    patch:
      summary: Update a todo item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                completed:
                  type: boolean
      responses:
        "200":
          description: Updated todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          description: Todo not found
    delete:
      summary: Delete a todo item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Todo deleted
        "404":
          description: Todo not found

  # ── Settings ──
...
components:
  schemas:
...
    Message:
      type: object
      required: [id, conversation_id, seq, sender_role, content]
      properties:
        id:
          type: string
        conversation_id:
          type: string
        seq:
          type: integer
        sender_role:
          type: string
          enum: [user, assistant, system, tool]
        sender_id:
          type: string
        content:
          type: string
        reason_content:
          type: string
        metadata:
          type: object
          additionalProperties: true
        finish_reason:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        duration_ms:
          type: integer
          nullable: true
        token_prompt:
          type: integer
        token_completion:
          type: integer
        created_at:
          type: string
          format: date-time
        reply_to_seq:
          type: integer
        mentioned_members:
          type: array
          items:
            type: string
        tool_calling:
          type: object
          nullable: true
          properties:
            input:
              type: object
              additionalProperties: true
              description: Tool call input arguments
            output:
              type: string
              description: Tool call output result
        updated_at:
          type: string
          format: date-time
...
    HumanInTheLoop:
      type: object
      required: [id, conversation_id, question, answer_type, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        checkpoint_id:
          type: string
        interrupt_id:
          type: string
        question:
          type: string
        choices:
          type: array
          items:
            type: object
            required: [title]
            properties:
              title:
                type: string
              desc:
                type: string
        answer_type:
          type: string
          enum: [single, multi, text]
        answer:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, answered, expired]
        created_at:
          type: string
          format: date-time
    Todo:
      type: object
      required: [id, conversation_id, content, completed, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        content:
          type: string
        completed:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...